+++ pkg/node/node.go
@@ -4,6 +4,7 @@
 package node
 
 import (
+	"context"
 	"reflect"
 
 	"github.com/mariomac/pipes/pkg/node/internal/connect"
@@ -12,6 +13,11 @@ import (
 // StartFunc is a function that receives a writable channel as unique argument, and sends
 // value to that channel during an indefinite amount of time.
 type StartFunc[OUT any] func(out chan<- OUT)
+
+// StartFuncCtx is a StartFunc that also receives a context as a first argument. If the passed
+// context is cancelled via the ctx.Done() function, the implementer function should end,
+// so the cancel will be propagated to the later nodes.
+type StartFuncCtx[OUT any] func(ctx context.Context, out chan<- OUT)
 
 // MiddleFunc is a function that receives a readable channel as first argument,
 // and a writable channel as second argument.
@@ -48,7 +54,7 @@ type Receiver[IN any] interface {
 // An Start node must have at least one output node.
 type Start[OUT any] struct {
 	outs    []Receiver[OUT]
-	fun     StartFunc[OUT]
+	fun     StartFuncCtx[OUT]
 	outType reflect.Type
 }
 
@@ -126,6 +132,13 @@ func (m *Terminal[IN]) InType() reflect.Type {
 
 // AsStart wraps an StartFunc into an Start node.
 func AsStart[OUT any](fun StartFunc[OUT]) *Start[OUT] {
+	return AsStartCtx(func(_ context.Context, out chan<- OUT) {
+		fun(out)
+	})
+}
+
+// AsStartCtx wraps an StartFuncCtx into an Start node.
+func AsStartCtx[OUT any](fun StartFuncCtx[OUT]) *Start[OUT] {
 	var out OUT
 	return &Start[OUT]{
 		fun:     fun,
@@ -158,7 +171,16 @@ func AsTerminal[IN any](fun TerminalFunc[IN], opts ...Option) *Terminal[IN] {
 	}
 }
 
+// Start the function wrapped in the Start node. Either this method or StartCtx should be invoked
+// for all the start nodes of the same graph, so the graph can properly start and finish.
 func (i *Start[OUT]) Start() {
+	i.StartCtx(context.TODO())
+}
+
+// StartCtx starts the function wrapped in the Start node, allow passing a context that can be
+// used by the wrapped function. Either this method or Start should be invoked
+// for all the start nodes of the same graph, so the graph can properly start and finish.
+func (i *Start[OUT]) StartCtx(ctx context.Context) {
 	if len(i.outs) == 0 {
 		panic("Start node should have outputs")
 	}
@@ -171,7 +193,7 @@ func (i *Start[OUT]) Start() {
 	}
 	forker := connect.Fork(joiners...)
 	go func() {
-		i.fun(forker.Sender())
+		i.fun(ctx, forker.Sender())
 		forker.Close()
 	}()
 }